separated string implementations into separate test class.
