+++ Tests/AttributesTests/validators/ValidatorFactoryTests.swift
@@ -285,254 +285,4 @@ final class ValidationFactoryTests: XCTestCase {
         }
     }
 
-    /// Test alpha rule.
-    func testAlpha() throws {
-        let readPath = ReadOnlyPath(String.self)
-        let factory = ValidatorFactory<String>.required().alpha()
-        try factory.make(path: readPath).performValidation("abc")
-        try factory.make(path: readPath).performValidation("aBc")
-        try factory.make(path: readPath).performValidation("")
-        XCTAssertThrowsError(try factory.make(path: readPath).performValidation("abc1")) {
-            handleStrError($0, message: "Must be alphabetic.")
-        }
-    }
-
-    /// Test `alphaDash` rule.
-    func testAlphaDash() throws {
-        let readPath = ReadOnlyPath(String.self)
-        let factory = ValidatorFactory<String>.required().alphadash()
-        try factory.make(path: readPath).performValidation("abc")
-        try factory.make(path: readPath).performValidation("aBc")
-        try factory.make(path: readPath).performValidation("")
-        try factory.make(path: readPath).performValidation("abc123")
-        try factory.make(path: readPath).performValidation("123")
-        try factory.make(path: readPath).performValidation("abc_123")
-        try factory.make(path: readPath).performValidation("abc-123")
-        XCTAssertThrowsError(try factory.make(path: readPath).performValidation("abc*!")) {
-            handleStrError($0, message: "Must be alphabetic with underscores and dashes allowed.")
-        }
-    }
-
-    /// Test `alphaFirst` rule.
-    func testAlphaFirst() throws {
-        let readPath = ReadOnlyPath(String.self)
-        let factory = ValidatorFactory<String>.required().alphafirst()
-        try factory.make(path: readPath).performValidation("abc")
-        try factory.make(path: readPath).performValidation("aBc")
-        try factory.make(path: readPath).performValidation("abc123")
-        try factory.make(path: readPath).performValidation("abc_123")
-        try factory.make(path: readPath).performValidation("abc-123")
-        try factory.make(path: readPath).performValidation("")
-        try factory.make(path: readPath).performValidation("abc*!")
-        XCTAssertThrowsError(try factory.make(path: readPath).performValidation("*!abc")) {
-            handleStrError($0, message: "First Character must be alphabetic.")
-        }
-        XCTAssertThrowsError(try factory.make(path: readPath).performValidation("123")) {
-            handleStrError($0, message: "First Character must be alphabetic.")
-        }
-    }
-
-    /// Test `alphanumeric` rule.
-    func testAlphaNumeric() throws {
-        let readPath = ReadOnlyPath(String.self)
-        let factory = ValidatorFactory<String>.required()
-        try doTest(
-            accepted: [
-                "123",
-                "abc123",
-                "aBc",
-                "",
-                "abc"
-            ],
-            errorsValues: [
-                "abc*!",
-                "*!abc",
-                "abc-123",
-                "abc_123"
-            ],
-            path: factory.alphanumeric().make(path: readPath),
-            message: "Must be alphanumeric."
-        )
-    }
-
-    /// Test `alphaUnderscore` rule.
-    func testAlphaUnderscore() throws {
-        let readPath = ReadOnlyPath(String.self)
-        let factory = ValidatorFactory<String>.required()
-        try doTest(
-            accepted: [
-                "123",
-                "abc123",
-                "aBc",
-                "",
-                "abc",
-                "abc_123"
-            ],
-            errorsValues: [
-                "abc*!",
-                "*!abc",
-                "abc-123"
-            ],
-            path: factory.alphaunderscore().make(path: readPath),
-            message: "Must be alphabetic with underscores allowed."
-        )
-    }
-
-    /// Test `alphaunderscorefirst` rule.
-    func testAlphaUnderscoreFirst() throws {
-        let readPath = ReadOnlyPath(String.self)
-        let factory = ValidatorFactory<String>.required()
-        try doTest(
-            accepted: [
-                "abc123",
-                "aBc",
-                "",
-                "abc",
-                "abc_123",
-                "_abc123*!",
-                "abc*!",
-                "abc-123"
-            ],
-            errorsValues: [
-                "123",
-                "*!abc"
-            ],
-            path: factory.alphaunderscorefirst().make(path: readPath),
-            message: "First Character must be alphabetic or an underscore."
-        )
-    }
-
-    /// Test `numeric` rule.
-    func testNumeric() throws {
-        let readPath = ReadOnlyPath(String.self)
-        let factory = ValidatorFactory<String>.required()
-        try doTest(
-            accepted: [
-                "123",
-                ""
-            ],
-            errorsValues: [
-                "abc123",
-                "aBc",
-                "abc",
-                "abc_123",
-                "_abc123*!",
-                "abc*!",
-                "abc-123",
-                "*!abc"
-            ],
-            path: factory.numeric().make(path: readPath),
-            message: "Must be numeric."
-        )
-    }
-
-    // swiftlint:disable inclusive_language
-
-    /// Test blacklist.
-    func testBlacklist() throws {
-        let banned: Set<String> = [
-            "abc123",
-            "aBc",
-            "",
-            "abc",
-            "abc_123",
-            "_abc123*!",
-            "abc*!",
-            "abc-123"
-        ]
-        let readPath = ReadOnlyPath(String.self)
-        let factory = ValidatorFactory<String>.required().blacklist(banned)
-        try factory.make(path: readPath).performValidation("123")
-        try factory.make(path: readPath).performValidation("!*abc")
-        XCTAssertThrowsError(try factory.make(path: readPath).performValidation("abc*!")) {
-            handleStrError($0, message: "abc*! is a banned word.")
-        }
-    }
-
-    /// Test `whitelist` rule.
-    func testWhitelist() throws {
-        let allowed: Set<String> = [
-            "abc123",
-            "aBc",
-            "",
-            "abc",
-            "abc_123",
-            "_abc123*!",
-            "abc*!",
-            "abc-123"
-        ]
-        let readPath = ReadOnlyPath(String.self)
-        let factory = ValidatorFactory<String>.required().whitelist(allowed)
-        try factory.make(path: readPath).performValidation("")
-        try factory.make(path: readPath).performValidation("abc*!")
-        XCTAssertThrowsError(try factory.make(path: readPath).performValidation("*!abc")) {
-            handleStrError(
-                $0, message: "*!abc is not valid, you must use pre-existing words. Candidates: \(allowed)"
-            )
-        }
-    }
-
-    /// Test `greyList` rule.
-    func testGreyList() throws {
-        let allowed: Set<String> = [
-            "abc123",
-            "aBc",
-            "abc",
-            "abc_123",
-            "_abc123*!",
-            "abc*!",
-            "abc-123",
-            "123"
-        ]
-        let readPath = ReadOnlyPath(String.self)
-        let factory = ValidatorFactory<String>.required().greyList(allowed)
-        try factory.make(path: readPath).performValidation("abcd")
-        try factory.make(path: readPath).performValidation("1234")
-        XCTAssertThrowsError(try factory.make(path: readPath).performValidation("xyz")) {
-            handleStrError(
-                $0, message: "xyz is not valid, it must contain pre-existing words. Candidates: \(allowed)"
-            )
-        }
-    }
-
-    // swiftlint:enable inclusive_language
-
-    /// Perform the test function.
-    /// - Parameters:
-    ///   - accepted: The values that pass validation.
-    ///   - errorsValues: The values that fail validation.
-    ///   - path: The validator performing the validation.
-    ///   - message: The error message when failing validation.
-    private func doTest(
-        accepted: [String],
-        errorsValues: [String],
-        path: AnyValidator<String>,
-        message: String
-    ) throws {
-        try accepted.forEach {
-            try path.performValidation($0)
-        }
-        try errorsValues.forEach { val in
-            XCTAssertThrowsError(try path.performValidation(val)) {
-                handleStrError($0, message: message)
-            }
-        }
-    }
-
-    /// Handle the case where the validator throws a validation error during a string validation.
-    /// - Parameters:
-    ///   - error: The abstract Error.
-    ///   - message: The message contained within the ``ValidationError``.
-    ///   - path: The path contained within the ``ValidationError``
-    private func handleStrError(
-        _ error: Error, message: String, path: AnyPath<String> = AnyPath(ReadOnlyPath(String.self))
-    ) {
-        guard let error = error as? ValidationError<String> else {
-            XCTFail("Incorrect error thrown.")
-            return
-        }
-        XCTAssertEqual(error.message, message)
-        XCTAssertEqual(error.path, path)
-    }
-
 }
